refactor: reorganize and update imports in `GetUserService` and related components
- Moved `DTO`, `Validators`, and `Query` files to appropriate `Application` and `Domain` subdirectories.
- Updated import paths across affected services, handlers, and view models.
