+++ user-service/src/Application/ViewModels/GetUserViewModel.ts
