+++ user-service/src/Domain/Queries/Handlers/GetUserQueryHandler.ts
@@ -1,9 +1,9 @@
 import {GetUserQuery} from "../QueryObject/GetUserQuery.js";
 import {UserRepository} from "../../../Infrastructure/Persistence/Repositories/Concrete/UserRepository.js";
 import {NotificationError} from "../../../Shared/Errors/NotificationError.js";
-import {GetUserDTO} from "../../../Domain/DTO/Query/GetUserDTO.js";
+import {GetUserDTO} from "../../../Application/DTO/ToQuery/GetUserDTO.js";
 import {BaseHandlerQuery} from "./BaseHandlerQuery.js";
-import {GetUserQueryDTO} from "../../../Domain/QueryDTO/GetUserQueryDTO.js";
+import {GetUserQueryDTO} from "../../QueryDTO/GetUserQueryDTO.js";
 
 export class GetUserQueryHandler implements BaseHandlerQuery<GetUserQuery | null>
 {
