+++ user-service/src/Domain/Queries/Validators/GetUserQueryValidator.ts
